refactor: Upload Package Component -> File Package로 이동
