+++ src/main/java/com/kgu/studywithme/file/utils/validator/FileExistsConstraint.java
@@ -1,4 +1,4 @@
-package com.kgu.studywithme.upload.utils.validator;
+package com.kgu.studywithme.file.utils.validator;
 
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
@@ -10,9 +10,9 @@ import java.lang.annotation.Target;
 
 @Target({ElementType.PARAMETER, ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = ImageUploadCategoryConstraintValidator.class)
-public @interface ImageUploadCategoryConstraint {
-    String message() default "잘못된 요청입니다.";
+@Constraint(validatedBy = FileExistsConstraintValidator.class)
+public @interface FileExistsConstraint {
+    String message() default "파일이 전송되지 않았습니다.";
 
     Class<?>[] groups() default {};
