+++ src/main/java/com/kgu/studywithme/file/utils/validator/ImageUploadCategoryConstraintValidator.java
@@ -1,4 +1,4 @@
-package com.kgu.studywithme.upload.utils.validator;
+package com.kgu.studywithme.file.utils.validator;
 
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
