--- src/main/java/com/kgu/studywithme/upload/presentation/dto/request/ImageUploadRequest.java
@@ -1,16 +1,0 @@
-package com.kgu.studywithme.upload.presentation.dto.request;
-
-import com.kgu.studywithme.upload.utils.validator.FileExistsConstraint;
-import com.kgu.studywithme.upload.utils.validator.ImageExtensionConstraint;
-import com.kgu.studywithme.upload.utils.validator.ImageUploadCategoryConstraint;
-import org.springframework.web.multipart.MultipartFile;
-
-public record ImageUploadRequest(
-        @ImageUploadCategoryConstraint
-        String type,
-
-        @FileExistsConstraint
-        @ImageExtensionConstraint
-        MultipartFile file
-) {
-}
