+++ src/test/java/com/kgu/studywithme/acceptance/file/FileUploadAcceptanceFixture.java
@@ -1,4 +1,4 @@
-package com.kgu.studywithme.acceptance.upload;
+package com.kgu.studywithme.acceptance.file;
 
 import io.restassured.response.ValidatableResponse;
 import org.springframework.web.util.UriComponentsBuilder;
@@ -8,7 +8,7 @@ import java.util.Map;
 
 import static com.kgu.studywithme.acceptance.CommonRequestFixture.multipartRequest;
 
-public class UploadAcceptanceFixture {
+public class FileUploadAcceptanceFixture {
     public static ValidatableResponse 스터디_주차_글_내부_이미지를_업로드한다(final String accessToken) {
         final String uri = UriComponentsBuilder
                 .fromPath("/api/image")
